// src/app/my_core/models/user.ts
export class User {
    id?: string;
    // username: string;
    password?: string;
    email: string;
    roles?: string[];
    pic?: string;
    fullname: string;
    phone: string;
    phone4?: string;
    phone1?: string;
    phone2?: string;
    phone3?: string;
    address?: string;
    cmt?: string;
    ngaycap?: string;
    noicap?: string;
    code?: string;
    list4personal?: string;
    idref: string;
}